+++ bricks/primehub_h5/manifest.py
@@ -1,0 +1,4 @@
+# Common modules
+include("../_common/manifest.py")
+
+freeze_as_mpy("../primehub_f4/modules", "_imu_calibrate.py")