+++ lib/pbio/drv/adc/adc_stm32_hal.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2019-2020 The Pybricks Authors
+// Copyright (c) 2019-2026 The Pybricks Authors
 
 // Platform should override HAL_ADC_MspInit() to configure pin mux and ADC
 // channel mapping. Example:
@@ -33,11 +33,20 @@
 
 #include STM32_HAL_H
 
+#if defined(STM32H5)
+#include <stm32h5xx_ll_dma.h>
+#endif
+
 #define PBDRV_ADC_PERIOD_MS 10  // polling period in milliseconds
 
 static TIM_HandleTypeDef pbdrv_adc_htim;
 static DMA_HandleTypeDef pbdrv_adc_hdma;
 static ADC_HandleTypeDef pbdrv_adc_hadc;
+
+#if defined(STM32H5)
+static DMA_NodeTypeDef pbdrv_adc_dma_node;
+static DMA_QListTypeDef pbdrv_adc_dma_queue;
+#endif
 
 static uint32_t pbdrv_adc_dma_buffer[PBDRV_CONFIG_ADC_STM32_HAL_ADC_NUM_CHANNELS];
 static uint32_t pbdrv_adc_error_count;
@@ -93,12 +102,75 @@ void pbdrv_adc_init(void) {
 
     // using DMA
 
+    pbdrv_adc_hadc.Instance = PBDRV_CONFIG_ADC_STM32_HAL_ADC_INSTANCE;
+
     pbdrv_adc_hdma.Instance = PBDRV_CONFIG_ADC_STM32_HAL_DMA_INSTANCE;
-    #ifdef STM32L4
+    #if defined(STM32L4) || defined(STM32H5)
     pbdrv_adc_hdma.Init.Request = PBDRV_CONFIG_ADC_STM32_HAL_DMA_REQUEST;
     #else
     pbdrv_adc_hdma.Init.Channel = PBDRV_CONFIG_ADC_STM32_HAL_DMA_CHANNEL;
     #endif
+    #if defined(STM32H5)
+    pbdrv_adc_hdma.InitLinkedList.Priority = DMA_LOW_PRIORITY_LOW_WEIGHT;
+    pbdrv_adc_hdma.InitLinkedList.LinkStepMode = DMA_LSM_FULL_EXECUTION;
+    pbdrv_adc_hdma.InitLinkedList.LinkAllocatedPort = DMA_LINK_ALLOCATED_PORT0;
+    pbdrv_adc_hdma.InitLinkedList.TransferEventMode = DMA_TCEM_BLOCK_TRANSFER;
+    pbdrv_adc_hdma.InitLinkedList.LinkedListMode = DMA_LINKEDLIST_CIRCULAR;
+
+    if (HAL_DMAEx_List_Init(&pbdrv_adc_hdma) != HAL_OK) {
+        pbdrv_adc_error_count++;
+        pbdrv_adc_last_error = pbdrv_adc_hdma.ErrorCode;
+        return;
+    }
+
+    DMA_NodeConfTypeDef dma_node_config = { };
+    dma_node_config.NodeType = DMA_GPDMA_LINEAR_NODE;
+    dma_node_config.Init.Request = PBDRV_CONFIG_ADC_STM32_HAL_DMA_REQUEST;
+    dma_node_config.Init.BlkHWRequest = DMA_BREQ_SINGLE_BURST;
+    dma_node_config.Init.Direction = DMA_PERIPH_TO_MEMORY;
+    dma_node_config.Init.SrcInc = DMA_SINC_FIXED;
+    dma_node_config.Init.DestInc = DMA_DINC_INCREMENTED;
+    dma_node_config.Init.SrcDataWidth = DMA_SRC_DATAWIDTH_WORD;
+    dma_node_config.Init.DestDataWidth = DMA_DEST_DATAWIDTH_WORD;
+    dma_node_config.Init.Priority = DMA_LOW_PRIORITY_LOW_WEIGHT;
+    dma_node_config.Init.SrcBurstLength = 1;
+    dma_node_config.Init.DestBurstLength = 1;
+    dma_node_config.Init.TransferAllocatedPort = DMA_SRC_ALLOCATED_PORT0 | DMA_DEST_ALLOCATED_PORT0;
+    dma_node_config.Init.TransferEventMode = DMA_TCEM_BLOCK_TRANSFER;
+    dma_node_config.Init.Mode = DMA_NORMAL;
+    dma_node_config.DataHandlingConfig.DataExchange = DMA_EXCHANGE_NONE;
+    dma_node_config.DataHandlingConfig.DataAlignment = DMA_DATA_RIGHTALIGN_ZEROPADDED;
+    dma_node_config.TriggerConfig.TriggerMode = DMA_TRIGM_BLOCK_TRANSFER;
+    dma_node_config.TriggerConfig.TriggerPolarity = DMA_TRIG_POLARITY_MASKED;
+    dma_node_config.TriggerConfig.TriggerSelection = 0;
+    dma_node_config.SrcAddress = (uint32_t)&pbdrv_adc_hadc.Instance->DR;
+    dma_node_config.DstAddress = (uint32_t)pbdrv_adc_dma_buffer;
+    dma_node_config.DataSize = sizeof(pbdrv_adc_dma_buffer);
+
+    if (HAL_DMAEx_List_BuildNode(&dma_node_config, &pbdrv_adc_dma_node) != HAL_OK) {
+        pbdrv_adc_error_count++;
+        pbdrv_adc_last_error = pbdrv_adc_hdma.ErrorCode;
+        return;
+    }
+
+    if (HAL_DMAEx_List_InsertNode_Tail(&pbdrv_adc_dma_queue, &pbdrv_adc_dma_node) != HAL_OK) {
+        pbdrv_adc_error_count++;
+        pbdrv_adc_last_error = pbdrv_adc_dma_queue.ErrorCode;
+        return;
+    }
+
+    if (HAL_DMAEx_List_SetCircularMode(&pbdrv_adc_dma_queue) != HAL_OK) {
+        pbdrv_adc_error_count++;
+        pbdrv_adc_last_error = pbdrv_adc_dma_queue.ErrorCode;
+        return;
+    }
+
+    if (HAL_DMAEx_List_LinkQ(&pbdrv_adc_hdma, &pbdrv_adc_dma_queue) != HAL_OK) {
+        pbdrv_adc_error_count++;
+        pbdrv_adc_last_error = pbdrv_adc_hdma.ErrorCode;
+        return;
+    }
+    #else
     pbdrv_adc_hdma.Init.Direction = DMA_PERIPH_TO_MEMORY;
     pbdrv_adc_hdma.Init.PeriphInc = DMA_PINC_DISABLE;
     pbdrv_adc_hdma.Init.MemInc = DMA_MINC_ENABLE;
@@ -106,10 +178,9 @@ void pbdrv_adc_init(void) {
     pbdrv_adc_hdma.Init.MemDataAlignment = DMA_MDATAALIGN_WORD;
     pbdrv_adc_hdma.Init.Mode = DMA_CIRCULAR;
     pbdrv_adc_hdma.Init.Priority = DMA_PRIORITY_MEDIUM;
-
     HAL_DMA_Init(&pbdrv_adc_hdma);
+    #endif
 
-    pbdrv_adc_hadc.Instance = PBDRV_CONFIG_ADC_STM32_HAL_ADC_INSTANCE;
     pbdrv_adc_hadc.Init.ClockPrescaler = ADC_CLOCK_SYNC_PCLK_DIV4;
     pbdrv_adc_hadc.Init.Resolution = ADC_RESOLUTION_12B;
     pbdrv_adc_hadc.Init.DataAlign = ADC_DATAALIGN_RIGHT;
@@ -125,7 +196,7 @@ void pbdrv_adc_init(void) {
 
     HAL_ADC_Init(&pbdrv_adc_hadc);
 
-    #ifdef STM32L4
+    #if defined(STM32L4) || defined(STM32H5)
     HAL_ADCEx_Calibration_Start(&pbdrv_adc_hadc, ADC_SINGLE_ENDED);
     #endif