+++ lib/pbio/drv/pwm/pwm_tlc5955_stm32.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2020 The Pybricks Authors
+// Copyright (c) 2020-2026 The Pybricks Authors
 
 // PWM driver using TI TLC5955 LED driver connected to STM32 MCU via SPI.
 
@@ -226,10 +226,25 @@ void pbdrv_pwm_tlc5955_stm32_init(pbdrv_pwm_dev_t *devs) {
         gpio_init.Pin = pdata->lat_gpio_pin;
         gpio_init.Mode = GPIO_MODE_OUTPUT_PP;
         gpio_init.Pull = GPIO_NOPULL;
-        gpio_init.Speed = GPIO_SPEED_LOW;
+        gpio_init.Speed = GPIO_SPEED_FREQ_LOW;
         HAL_GPIO_Init(pdata->lat_gpio, &gpio_init);
 
         priv->hdma_rx.Instance = pdata->rx_dma;
+        #if defined(STM32H5)
+        priv->hdma_rx.Init.Request = pdata->rx_dma_req;
+        priv->hdma_rx.Init.BlkHWRequest = DMA_BREQ_SINGLE_BURST;
+        priv->hdma_rx.Init.Direction = DMA_PERIPH_TO_MEMORY;
+        priv->hdma_rx.Init.SrcInc = DMA_SINC_FIXED;
+        priv->hdma_rx.Init.DestInc = DMA_DINC_INCREMENTED;
+        priv->hdma_rx.Init.SrcDataWidth = DMA_SRC_DATAWIDTH_BYTE;
+        priv->hdma_rx.Init.DestDataWidth = DMA_DEST_DATAWIDTH_BYTE;
+        priv->hdma_rx.Init.Priority = DMA_LOW_PRIORITY_LOW_WEIGHT;
+        priv->hdma_rx.Init.SrcBurstLength = 1;
+        priv->hdma_rx.Init.DestBurstLength = 1;
+        priv->hdma_rx.Init.TransferAllocatedPort = DMA_SRC_ALLOCATED_PORT0 | DMA_DEST_ALLOCATED_PORT0;
+        priv->hdma_rx.Init.TransferEventMode = DMA_TCEM_BLOCK_TRANSFER;
+        priv->hdma_rx.Init.Mode = DMA_NORMAL;
+        #else
         priv->hdma_rx.Init.Channel = pdata->rx_dma_ch;
         priv->hdma_rx.Init.Direction = DMA_PERIPH_TO_MEMORY;
         priv->hdma_rx.Init.PeriphInc = DMA_PINC_DISABLE;
@@ -241,12 +256,28 @@ void pbdrv_pwm_tlc5955_stm32_init(pbdrv_pwm_dev_t *devs) {
         priv->hdma_rx.Init.FIFOMode = DMA_FIFOMODE_DISABLE;
         priv->hdma_rx.Init.MemBurst = DMA_MBURST_SINGLE;
         priv->hdma_rx.Init.PeriphBurst = DMA_PBURST_SINGLE;
+        #endif
         HAL_DMA_Init(&priv->hdma_rx);
 
         HAL_NVIC_SetPriority(pdata->rx_dma_irq, 7, 0);
         HAL_NVIC_EnableIRQ(pdata->rx_dma_irq);
 
         priv->hdma_tx.Instance = pdata->tx_dma;
+        #if defined(STM32H5)
+        priv->hdma_tx.Init.Request = pdata->tx_dma_req;
+        priv->hdma_tx.Init.BlkHWRequest = DMA_BREQ_SINGLE_BURST;
+        priv->hdma_tx.Init.Direction = DMA_MEMORY_TO_PERIPH;
+        priv->hdma_tx.Init.SrcInc = DMA_SINC_INCREMENTED;
+        priv->hdma_tx.Init.DestInc = DMA_DINC_FIXED;
+        priv->hdma_tx.Init.SrcDataWidth = DMA_SRC_DATAWIDTH_BYTE;
+        priv->hdma_tx.Init.DestDataWidth = DMA_DEST_DATAWIDTH_BYTE;
+        priv->hdma_tx.Init.Priority = DMA_LOW_PRIORITY_LOW_WEIGHT;
+        priv->hdma_tx.Init.SrcBurstLength = 1;
+        priv->hdma_tx.Init.DestBurstLength = 1;
+        priv->hdma_tx.Init.TransferAllocatedPort = DMA_SRC_ALLOCATED_PORT0 | DMA_DEST_ALLOCATED_PORT0;
+        priv->hdma_tx.Init.TransferEventMode = DMA_TCEM_BLOCK_TRANSFER;
+        priv->hdma_tx.Init.Mode = DMA_NORMAL;
+        #else
         priv->hdma_tx.Init.Channel = pdata->tx_dma_ch;
         priv->hdma_tx.Init.Direction = DMA_MEMORY_TO_PERIPH;
         priv->hdma_tx.Init.PeriphInc = DMA_PINC_DISABLE;
@@ -258,6 +289,7 @@ void pbdrv_pwm_tlc5955_stm32_init(pbdrv_pwm_dev_t *devs) {
         priv->hdma_tx.Init.FIFOMode = DMA_FIFOMODE_DISABLE;
         priv->hdma_tx.Init.MemBurst = DMA_MBURST_SINGLE;
         priv->hdma_tx.Init.PeriphBurst = DMA_PBURST_SINGLE;
+        #endif
         HAL_DMA_Init(&priv->hdma_tx);
 
         HAL_NVIC_SetPriority(pdata->tx_dma_irq, 7, 1);