+++ lib/pbio/platform/prime_hub_h5/pbsysconfig.h
@@ -1,0 +1,31 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2020-2026 The Pybricks Authors
+
+#define PBSYS_CONFIG_FEATURE_BUILTIN_USER_PROGRAM_REPL             (1)
+#define PBSYS_CONFIG_FEATURE_BUILTIN_USER_PROGRAM_PORT_VIEW        (1)
+#define PBSYS_CONFIG_FEATURE_BUILTIN_USER_PROGRAM_IMU_CALIBRATION  (0)
+#define PBSYS_CONFIG_FEATURE_PROGRAM_FORMAT_MULTI_MPY_V6           (1)
+#define PBSYS_CONFIG_FEATURE_PROGRAM_FORMAT_MULTI_MPY_V6_3_NATIVE  (1)
+#define PBSYS_CONFIG_BATTERY                        (1)
+#define PBSYS_CONFIG_BATTERY_CHARGER                (1)
+#define PBSYS_CONFIG_HMI                            (1)
+#define PBSYS_CONFIG_HMI_STOP_BUTTON                (1 << 5) // center
+#define PBSYS_CONFIG_HMI_NUM_SLOTS                  (5)
+#define PBSYS_CONFIG_HMI_PUP                        (1)
+#define PBSYS_CONFIG_HMI_PUP_LIGHT_MATRIX_INDEX     (0)
+#define PBSYS_CONFIG_HMI_PUP_BLUETOOTH_BUTTON       (1 << 9) // right up
+#define PBSYS_CONFIG_HMI_PUP_LEFT_RIGHT_ENABLE      (1)
+#define PBSYS_CONFIG_HUB_LIGHT_MATRIX               (1)
+#define PBSYS_CONFIG_HUB_LIGHT_MATRIX_LED_ARRAY     (1)
+#define PBSYS_CONFIG_HOST                           (1)
+#define PBSYS_CONFIG_HOST_STDIN_BUF_SIZE            (64)
+#define PBSYS_CONFIG_MAIN                           (1)
+#define PBSYS_CONFIG_STORAGE                        (1)
+#define PBSYS_CONFIG_STORAGE_NUM_SLOTS              (5)
+#define PBSYS_CONFIG_STORAGE_USER_DATA_SIZE         (512)
+#define PBSYS_CONFIG_STATUS_LIGHT                   (1)
+#define PBSYS_CONFIG_STATUS_LIGHT_BATTERY           (1)
+#define PBSYS_CONFIG_STATUS_LIGHT_BLUETOOTH         (1)
+#define PBSYS_CONFIG_STATUS_LIGHT_STATE_ANIMATIONS  (0)
+#define PBSYS_CONFIG_USER_PROGRAM                   (1)
+#define PBSYS_CONFIG_PROGRAM_STOP                   (1)